+++ pylama.ini
@@ -1,0 +1,12 @@
+[pylama]
+async = 1
+linters = pyflakes,mccabe,pydocstyle,pylint,isort
+ignore = W0123,W0613,W0703,W1203,E0401,E0611,C0103,C0200,C0415,C901,R0902,R0914,R0903,R1732,W0107,R0913,D104,D105,D212,D203,D204,D213,D215,D400,D401,D404,D406,D407,D408,D409,D413
+# All warnings after D203 are to adopt the Google docstring style
+
+[pylama:pylint]
+max-line-length = 119
+good-names = w,h,k,v,x,y,a,b,i,j,f
+extension-pkg-whitelist = cv2
+max-branches = 40
+max-statements = 120
