+++ requirements.txt
@@ -1,0 +1,8 @@
+# Automatically generated by https://github.com/damnever/pigar.
+
+Pillow >= 8.1.2
+logzero >= 1.7.0
+matplotlib >= 3.4.0
+numpy >= 1.19.5
+opencv_python >= 4.5.1.48
+tqdm >= 4.59.0
